web/requirements.txt
flask
flask-login
psycopg2-binary